Move modes and remove matrix parameters
