+++ Scripts/assignment/datatypes/transit.py
@@ -1,61 +1,53 @@
-from __future__ import annotations
-from typing import Any, Dict, Union
 import parameters.assignment as param
+from assignment.datatypes.assignment_mode import AssignmentMode
 from assignment.datatypes.journey_level import JourneyLevel
 
 
-class TransitSpecification:
-    """
-    Transit assignment specification.
-    
-    Two journey levels are added at a later stage.
-    At the second level an extra boarding penalty is implemented,
-    hence a transfer penalty. Waiting time length is also different. 
-    Walk only trips are not allowed.
+class TransitMode(AssignmentMode):
+    def __init__(self, day_scenario, *args):
+        AssignmentMode.__init__(self, *args)
+        self.vot_inv = param.vot_inv[param.vot_classes[self.name]]
+        self.num_board = self._create_matrix("num_board")
+        self.gen_cost = self._create_matrix("gen_cost")
+        self.inv_cost = self._create_matrix("inv_cost")
+        self.board_cost = self._create_matrix("board_cost")
+        self.transit_matrices = {}
+        for subset, parts in param.transit_impedance_matrices.items():
+            self.transit_matrices[subset] = {}
+            for mtx_type, longer_name in parts.items():
+                self.transit_matrices[subset][longer_name] = self._create_matrix(mtx_type)
+        self.specify(day_scenario)
 
-    Parameters
-    ----------
-    transit_class : str
-        Name of transit class (transit_work/transit_leisure/...)
-    segment_results : dict
-        key : str
-            Segment result (transit_volumes/...)
-        value : str
-            Extra attribute name (@transit_work_vol_aht/...)
-    park_and_ride_results : str or False (optional)
-        Extra attribute name for park-and-ride aux volume if
-        this is park-and-ride assignment, else False
-    headway_attribute : str
-        Line attribute where headway is stored
-    emme_matrices : dict
-        key : str
-            Impedance type (time/cost/dist/...)
-        value : str
-            Emme matrix id
-    count_zone_boardings : bool (optional)
-        Whether assignment is performed only to count fare zone boardings
-    """
-    def __init__(self, 
-                 transit_class: str,
-                 segment_results: Dict[str,str],
-                 park_and_ride_results: Union[str, bool],
-                 headway_attribute: str,
-                 emme_matrices: Dict[str, Union[str, Dict[str, str]]], 
-                 count_zone_boardings: bool = False):
+    def specify(self, day_scenario):
+        self.segment_results = {}
+        for emme_scenario, tp in (
+                (self.emme_scenario, self.time_period), (day_scenario, "vrk")):
+            for res, attr in param.segment_results.items():
+                attr_name = f"@{self.name[:11]}_{attr}_{tp}"
+                self.segment_results[res] = attr_name
+                self.emme_project.create_extra_attribute(
+                    "TRANSIT_SEGMENT", attr_name,
+                    f"{self.name} {res}", overwrite=True,
+                    scenario=emme_scenario)
+                if res != "transit_volumes":
+                    self.emme_project.create_extra_attribute(
+                        "NODE", f"@{self.name[:10]}n_{attr}_{tp}",
+                        f"{self.name} {res}", overwrite=True,
+                        scenario=emme_scenario)
         no_penalty = dict.fromkeys(["at_nodes", "on_lines", "on_segments"])
         no_penalty["global"] = {
-            "penalty": 0, 
+            "penalty": 0,
             "perception_factor": 1,
         }
         modes = (param.local_transit_modes + param.aux_modes
-                 + param.long_dist_transit_modes[transit_class])
-        self.transit_spec: Dict[str, Any] = {
+                 + param.long_dist_transit_modes[self.name])
+        self.transit_spec = {
             "type": "EXTENDED_TRANSIT_ASSIGNMENT",
             "modes": modes,
-            "demand": emme_matrices["demand"],
+            "demand": self.demand.id,
             "waiting_time": {
                 "headway_fraction": 1,
-                "effective_headways": headway_attribute,
+                "effective_headways": param.effective_headway_attr,
                 "spread_factor": 1,
                 "perception_factor": 1,
             },
@@ -63,7 +55,7 @@ class TransitSpecification:
                 "global": None,
                 "at_nodes": None,
                 "on_lines": {
-                    "penalty": param.boarding_penalty_attr + transit_class,
+                    "penalty": param.boarding_penalty_attr + self.name,
                     "perception_factor": 1
                 },
                 "on_segments": param.extra_waiting_time,
@@ -76,8 +68,7 @@ class TransitSpecification:
             },
             "in_vehicle_cost": {
                 "penalty": param.line_penalty_attr,
-                "perception_factor": (param.vot_inv[param.vot_classes[
-                    transit_class]]),
+                "perception_factor": self.vot_inv,
             },
             "aux_transit_time": param.aux_transit_time,
             "flow_distribution_at_origins": {
@@ -92,29 +83,37 @@ class TransitSpecification:
             "journey_levels": None,
             "performance_settings": param.performance_settings,
         }
-        if park_and_ride_results:
+        if self.name in param.park_and_ride_classes:
+            self.park_and_ride_results = f"@{self.name[4:]}_aux"
+            self.emme_project.create_extra_attribute(
+                "LINK", self.park_and_ride_results, self.name,
+                overwrite=True, scenario=self.emme_scenario)
             self.transit_spec["modes"].append(param.park_and_ride_mode)
             self.transit_spec["results"] = {
                 "aux_transit_volumes_by_mode": [{
                     "mode": param.park_and_ride_mode,
-                    "volume": park_and_ride_results,
+                    "volume": self.park_and_ride_results,
                 }],
             }
+        else:
+            self.park_and_ride_results = False
         self.transit_spec["journey_levels"] = [JourneyLevel(
-                level, transit_class, headway_attribute, park_and_ride_results,
-                count_zone_boardings).spec
+                level, self.name, self.park_and_ride_results).spec
             for level in range(6)]
         self.ntw_results_spec = {
             "type": "EXTENDED_TRANSIT_NETWORK_RESULTS",
-            "analyzed_demand": emme_matrices["demand"],
-            "on_segments": segment_results,
+            "analyzed_demand": self.demand.id,
+            "on_segments": self.segment_results,
             }
         subset = "by_mode_subset"
         self.transit_result_spec = {
             "type": "EXTENDED_TRANSIT_MATRIX_RESULTS",
             subset: {
                 "modes": modes,
-                "distance": emme_matrices["dist"],
+                "total_impedance": self.gen_cost.id,
+                "actual_in_vehicle_costs": self.inv_cost.id,
+                "actual_total_boarding_costs": self.board_cost.id,
+                "avg_boardings": self.num_board.id,
             },
         }
         self.local_result_spec = {
@@ -123,14 +122,18 @@ class TransitSpecification:
                     "modes": param.local_transit_modes,
                 },
             }
-        if count_zone_boardings:
-            bcost = "actual_total_boarding_costs"
-            self.transit_result_spec[subset][bcost] = emme_matrices[bcost]
-        else:
-            self.transit_result_spec["total_impedance"] = emme_matrices["gen_cost"]
-            for trip_part, matrix_id in emme_matrices["total"].items():
-                self.transit_result_spec[trip_part] = matrix_id
-            for trip_part, matrix_id in emme_matrices[subset].items():
-                self.transit_result_spec[subset][trip_part] = matrix_id
-            for trip_part, matrix_id in emme_matrices["local"].items():
-                self.local_result_spec[subset][trip_part] = matrix_id
+        for trip_part, matrix in self.transit_matrices["total"].items():
+            self.transit_result_spec[trip_part] = matrix.id
+        for trip_part, matrix in self.transit_matrices[subset].items():
+            self.transit_result_spec[subset][trip_part] = matrix.id
+        for trip_part, matrix in self.transit_matrices["local"].items():
+            self.local_result_spec[subset][trip_part] = matrix.id
+
+    def get_matrices(self):
+        transfer_penalty = ((self.num_board.data > 0)
+                            * param.transfer_penalty[self.name])
+        cost = self.inv_cost.data + self.board_cost.data
+        time = self.gen_cost.data - self.vot_inv*cost - transfer_penalty
+        self._release_matrices()
+        time[cost > 999999] = 999999
+        return {"time": time, "cost": cost}
